New command line setup for repl
